Add crate for platform testing
Adds two new crates `uefi_test` and `crates\uefi_test_macro` to the
workspace. This crate provides macros and a dxe_core component for
collating and running tests on a physical or virtual platform.

The `uefi_test` attribute macro (`#[uefi_test]`) is used to define a
test that should be run. During build, all functions marked with this
attribute are gathered into a contiguous section of the binary by the
linker. At runtime, the `TestRunnerComponent` should be registered with
the dxe_core and will automatically run all tests.

Multiple configuration options exist both for the `uefi_test` attribute
macro and for the `TestRunnerComponent` to allow for a variety of test
customization and test execution customization. Please review the
documentation provided with this crate for more information.