dxe_core: Add EFI_MEMORY_MAP trace logging (#117)
## Description

This adds a new target for logging: efi_memory_map. When set to debug
level for a platform, the EFI_MEMORY_MAP will be pretty printed in the
log whenever it is created. Info logging was chosen over trace because
trace logging prints the file name before each line, which takes up too
much room for what this table needs to be easily read and it is
distracting with so many lines.

This commits also changes the MAT logic to use the memory descriptor
pretty print logic created in this commit, though it does not tie it to
the efi_memory_map logging target or any other, as the MAT is created
less often and has fewer members and is useful to be in the log by
default.

The column sizes and individual column member spacing was tested and
adjusted so it lined up visually, even though the numbers in the format
strings appear off.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested with printing enabled/disabled for target and seeing both memory
map and MAT and only MAT prints.

## Integration Instructions

In the platform's dxe core binary wrapper, set

```rust
uefi_logger::Format::Standard,
    &[
        ("efi_memory_map", log::LevelFilter::Debug),
    ],
```
to enable printing the EFI_MEMORY_MAP.