Makefile: Do not run coverage for `uefi_test_macro`
Update the coverage command in the MakeFile to test the entire workspace
but filter out the `uefi_test_macro` crate. This crate is a proc-macro
crate which seems to be having issues with the coverage tooling. I've
reached out to the maintainers of the `tarpaulin` tool to see if they
have any suggestions on how to resolve the issue.