allocator: Update ensure_capacity to call alloc/dealloc
There was a PR comment to use alloc/dealloc in the
implementation of ensure_capacity, which was implemented
and tested, but not pushed before the PR was merged.

This is the change that was requested in addition to the
other PR feedback of adding an extra comment empty line
before the function name.