if_<arch>! macro: Use doc feature flag, not doc cfg flag (#13)
## Description

The `if_<arch>!` macro currently uses the doc cfg flag to compile
any architecture specific items and documentation. This works fine for
generating documentation in the same repository but the doc cfg flag
does not cascade down to dependent crates. This means that any
crates depending on architecture specific items in these crates will not
exist and will error and warn with
`note: found an item that was configured out`

The original implementation of the if_<arch>! macro was to use the
doc feature flag to compile, but was incorrectly set to just "doc" (
the cfg flag). This commit updates it to use the doc feature flag.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Confirmed documentation still builds as expected, but that documentation
also builds in consumer crates

## Integration Instructions

consumer crates must set the "doc" feature flag when building
documentation.