image loader: Add Stack Guard for Loaded Images
Each image loaded in DxeRust gets a separate stack created
for it and it is executed using that stack with corosensei.

We implement our stack functionality that we pass to corosensei,
so no guard page is added by default to the stack. This commit
adds a guard page at the bottom of the stack to catch stack
overflows in drivers/applications.

This does not cover DXE core itself, that is a change that will
come later, today DXE core has its stack and stack guard set
up by DxeIpl. In the future, DXE Core will create its own
stack.