.github: Build mdbook in CI workflow
No tests of the mdbook are run in the CI workflow. They will only fail
after a PR is merged and the publish-mdbook job is run. This change adds
a mdbook build step the the basic_ci job in the CiWorkflow.yml file as
it does not need to be tested on both operating systems.