.github: Add cargo clippy to the CI workflow
Cargo clippy was not originally added to the CI workflow as it is
strongly opinionated and the suggested changes can change the logic of
the code. However, we still want to run clippy and have the developer
decide if the changes are appropriate or add the attribute tags to
the code to ignore the clippy suggestions.