uefi_test: cleanup macro usage (#144)
## Description

in the uefi_test macro, we originally manually expanded the code
generated by the `distributed_slice` macro from linkme, as we were under
the impression that the path to the linkme crate was hardcoded as
`::linkme`. After additional review of the linkme code (because it was
not mentioned in the linkme documentation), it was discovered that an
attribute macro is available to override the path to the linkme crate
(`#[linkme(crate = "path::to::linkme")]`).

This commit removes the manual expansion of the macro and replaces it
with the macro itself, using the path override attribute to specify the
path to the linkme crate. An additional benefit is that the tests can
now focus only on the static test case our macro is generating.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Verified that the code continues to compile and the tests continue to be
located.

## Integration Instructions

N/A