dxe_core: Resolved unused code warnings. (#172)
## Description

Within multiple modules that were previously standalone crates (
specifically, uefi_allocator, uefi_event, uefi_protocol_db, and
uefi_gcd), multiple functions were unused in the code base. However
these functions were marked as public and were part of the public
interface for their respective crates, so no warnings were generated as
this is typically expected.

After moving these crates inwards, they are no longer apart of the
crate's public interface. Due to this, the compiler correctly detects
and reports that they are unused. An interesting note, however, is that
the compiler does not detect this when building the crate itself, only
when building a binary that consumes this crate.

This commit resolves these warnings primarily by marking these functions
with the #[allow(dead_code)] attribute, as they are hypothetically
useful, and could start to be consumed in the future. A few of the
functions were configured to only exist for test builds as they are for
testing purposes only.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Passes CI, and qemu binaries continue to build

## Integration Instructions

N/A