dxe_core: GCD: Silently Fail Free Calls During EBS
The original commit that disallowed alloc/free during EBS accidently
went too far and debug_asserted and returned an error on all frees
during EBS. This was originally only intended for runtime memory
types, as those would change the OS memory map. This now blocks
platforms and so is fixed by silently failing all frees. The runtime
types will be debug_asserted and return an error, but that requires
more code reorganization, so is deferred for now and tracked in a
GH issue. See the comment in this commit for more details.