Test Improvements (#178)
## Description

With crate->module conversion, tests that used to be separate
compilation units and therefore with separate global state now share
more global state (for example, anything using a TplMutex is subject to
TplMutex global state around the boot services pointer). This PR adds
global state lock guards to all tests with global state to account for
this.

This PR also adds a catch/unwind to the global state guard to avoid
scenarios where a panic in a test causes the global state lock to get
poisoned, which causes false failures on every other test waiting on the
global state lock. The code now catches the panic, then allows the lock
to be released, and then re-throws the panic with `.unwrap()` on the
catch/unwind closure. This means that panics in a test are now properly
caught by the test harness without poisoning the global state lock.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Unit tests pass consistently. Several deadlocky/mutex fail type
intermittent issues are no longer observed.

## Integration Instructions

N/A - but when writing new tests, ensure that if they touch any kind of
global state that they use the global state lock wrapper.