Disable default features in x86_64 to avoid picking up dependency on nightly (#7)
## Description

x86_64 default features includes "nightly", which include changes that
don't build with current stable. This causes CI breakage when using the
recent 0.14.13 release. This change disables the default features of
x86_64 to avoid this dependency.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Local build to confirm that 0.14.13 release of x86_64 can be used
without breaking the build on stable toolchains.

## Integration Instructions

N/A