dxe_core: Check DXE Core Mem Allocation Module HOB GUID (#185)
## Description

Closes #182 

Currently the first Memory Allocation Module HOB is assumed to be the
DXE Core. To make identification more robust and ensure other
assumptions about the DXE Core GUID are correct, this change also checks
that a HOB is present with a DXE Core module name GUID that matches.

- [x] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- Cargo build and test
- QEMU boot with expected and unexpected DXE Core FFS GUID

## Integration Instructions

- Ensure that the DXE Core GUID used matches
[`uefi_sdk::guid::DXE_CORE`](https://github.com/pop-project/uefi-sdk/blob/128e0e2464750d527d3615153fc0b173390e568f/uefi_sdk/src/guid.rs#L24-L34)
(`23C9322F-2AF2-476A-BC4C-26BC88266C71`). For this reason, the change
has been marked as potentially breaking.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>