rust-tool-cache: use cargo binstall (#196)
## Description

cargo install has shown to be an unreliable way to compile and install
tools for any project that does not use the current latest version of
rust. This is because some tools allow for relaxed versions of their
dependencies, which will cause cargo to attempt to use the latest
version of the dependency, which may not be compatible with the 
version of rust being used in the repository.

This change switches to using cargo binstall, which simply downloads
the compiled tool, with a fallback to cargo install if the tool does
not support binstall.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated installing tools continues to work when cache fails.

## Integration Instructions

N/A