Add 32 bit Code Segment to GDT and Set All Selectors
The current GDT in uefi-core does not set a 32 bit code segment.
This will immediately crash any APs starting and attempting to
transition to protected mode. This patch adds a flat 32 bit code
segment.

The current GDT also does not set all the segment selectors. This
causes a crash in the bootloader when it attempts to switch back to
FW context. It uses the old DS value, which contained a valid offset
for the PEI GDT, but does not contain a valid offset for the DXE GDT.
This causes a GP fault. Because long mode does not use segmentation,
this is only an issue when the DS attempts to be written to again.

These two issues did not show up when CpuDxe was being used because
it creates its own GDT before the APs are started or the bootloader
is launched.

This commit also punts two assumptions the core has: that the GDT
and IDT are allocated below 4GB. This will always be the case for
a 32 bit PEI, but is not necessarily so for a 64 bit PEI. This
commit adds logic to panic if the GDT or IDT is above 4GB, because
the APs will fault when running through protected mode. A follow
up should be done to simply relocate these two structures to below
4GB if DXE Core was allocated above 4GB.