Return Correct Result From GCD Functions
There was no semicolon after the map change callback functions,
meaning that we would return the wrong result if the callback
succeeded. In practice, this would rarely matter, but if the code
is shifted so that other code should execute after the callback,
it will not do so (as was the case in AllocateMemorySpace, fixed
previously).