Adding CPU arch support for AArch64 (#203)
## Description

This change adds the CPU Arch protocol and hardware interrupt protocol
for AArch64.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This change was tested on QEMU SBSA platform. With a CpuDxe driver
modified to only support memory attribute setting, the system can boot
to UEFI shell.

## Integration Instructions

`InterruptBases` needs to be initiated before handing off the core init.