dxe_core: Use Type-State pattern (#224)
Convert the dxe_core Core struct to use the Type-State pattern to
prevent allocations for the short period of time before the global
allocator is fully initialized instead of having two completely separate
structs for the two different states.

The Type-State pattern is a way to gate functionality to certain states
by using the type system and phantom data to implement functions for
certain states and prevent them, at compile time, from being used in
other states.

## Description

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

CI passes and the platform builds with no integration steps

## Integration Instructions

N/A