storage: Add pub function to lock all config (#21)
## Description

Adds a public function to the Storage object that allows the caller to
lock all config values so that no further changes can be made.

Due to the fact that no component that needs `Config<T>` will be
executed while the raw config is unlocked, this function will be used by
the DXE Core to lock all configuration values before attempting a second
round of dispatch. The flow will be:

1. Try dispatch all components (loop until all are executed or the count
has not changed)
2. Lock all configs
3. Try dispatch all remaining components (loop until all are executed or
the count has not changed

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

N/A