storage: Add ConfigRaw container for untyped config (#22)
## Description

This adds a new container type, ConfigRaw, which is a container for the
previous tuple object `(bool, Box<dyn Any>)` This update is non
functional, but provides a more descriptive type and easier usability
and readability for developers.

One actual benefit to this change is the ability to prevent public users
from being able to unlock a config object. Config objects should only be
unlocked once when a component is initialized that requires ConfigMut
access to the config object.

This was a suggestion offline from @cfernald which I think is a great
addition.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

N/A