CiWorkflow: Remove GH_PAT requirement (#227)
## Description

with the move from direct repository access for dependencies to using a
private registry, the GH_PAT is no longer needed for CI. Instead, the
CRATES_IO_TOKEN is used to authenticate with the private registry.

This commit updates the required secret to be CRATES_IO_TOKEN instead of
GH_PAT.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

CI passes

## Integration Instructions

Consuming repositories should add the CRATES_IO_TOKEN secret to their
list of secrets, and remove the GH_PAT secret.