component: Update default state of config to locked (#26)
## Description

The default state of a config when created should be locked. A config
should be updated to `unlocked` only when a component that uses
`ConfigMut` is registered. This change reduces the amount of overall
dispatch iterations necessary to execute all components with `Config<T>`
as config that is never mutated is locked by default, allowing
components that require `Config<T>` to be executed on the first dispatch
iteration instead of only once the dispatcher locks all configs.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Ensured components that only require `Config<T>` execute during the
first iteration when no other components also require `ConfigMut<T>` for
the same `T`.

## Integration Instructions

N/A