require uefi_sdk ^2 (#239)
## Description

Updates uefi-dxe-core to consume uefi_sdk ^2 and makes the appropriate
changes.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Continues to compile and pass CI

## Integration Instructions

Platforms that wish to consume this version of uefi-dxe-core or greater
will need to update their uefi_sdk to ^2 and perform any integrations as
specified in
[uefi_sdk@2.0.0](https://github.com/OpenDevicePartnership/uefi-sdk/releases/tag/v2.0.0),
but no integrations are necessary from the viewpoint of this repository.