component: Drop &HobList param support (#32)
## Description

In our efforts to keep the coupling between EDKII and the pure rust DXE
Core as low as possible, we have decided to remove the &HobList object
from the mu_pi crate as a possible injectable parameter. As there are
currently no components that use it, I will not be marking this as a
breaking change.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

N/A