Fix an issue where device path handle wasn't set for images not loaded from an FV. (#246)
## Description

Images loaded from a file system (such as an OS loader application) did
not have a correct device_handle in the loaded_image structure if a
source buffer was provided to core_load_image (i.e. they are loaded from
memory but have a non-memory file_path device path input). This was
working prior to the changes from #240 because the simple implementation
prior to that was always furnishing some device_handle.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Verified that images loaded from disk with a source buffer now have a
device_handle in the loaded_image.

## Integration Instructions

N/A