uefi_debugger: switch no_alloc to alloc feature (#72)
## Description

Features are expected to be "additive", not "subtractive", i.e. by
enabling a feature, you add more functionality. This is why the standard
for enabling / disabling allocations in public rust crates is to have an
"alloc" feature rather a "no_alloc" features.

This is somewhat argumentative in the scenario of the UEFI Debugger
because by adding the "alloc" feature, we also _lose_ the ability to
debug before allocations are available. However to stick with industry
standards, we switch to the "alloc" feature.

To prevent this from being a breaking change, we add the "alloc" feature
to the default feature set.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

Users must now disable default features instead of enabling "no_alloc"
feature. to debug prior to allocations being available