Separate internal and public advanced logger protocol structures  (#264)
## Description

Splits the existing protocol definition into a pure representation of
the C struct for general use and a wrapper for internal use that carries
the rust reference and generics. This simplifies the registration model
as well as removes the need for the consumer to know the generic type.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on Q35 with integration test.

## Integration Instructions

N/A