Support Unit Testing and Code Coverage for Firmware Volume (#250)
## Description

This is for supporting the Firmware Volume Unit Test Cases and Coverage.

Please note that Since all the Error Types need to be covered, the data
structures are kept common for all valid and invalid cases.

The current coverage % is increased to 82.56%. The original with out
changes were 24%.
This also exposed some bugs underlying too which I am planning to raise
bugs for.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

cargo make coverage and checking the coverage report.

## Integration Instructions

N/A

---------

Co-authored-by: Rathi Kartheek <rkartheek@microsoft.com>