Uefi-Cpu: Paging: Revert back null implementation (#80)
## Description

Because the core is arch independent. The GCD expects "an"
implementation of `create_cpu_paging()` to the present always in
uefi-cpu crate. During my cleanup in #75 I removed the null
implementation, and my local testing did not catch this because the
cargo.lock fooled me. This is now causing the pipelines for
`uefi-dxe-core` to fail when running clippy with
`--all-targets/--all-features` enabled.

- [x] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Built successfully.

## Integration Instructions

NA

Signed-off-by: Vineel Kovvuri[MSFT] <vineelko@microsoft.com>