component: add `Service` Parameter type (#36)
## Description

Adds a new parameter type that can be used in a component, which defines
a `Service` that offers abstracted functionality to the component. That
is to say, the component does not need to know the backing
implementation of the service, only the methods that it can call
(defined by a trait). e.g. `Service<dyn MyTrait>`

The `Service` parameter can also be used against a concrete type if
desired. The only use case for this is to allow the Service to also be
registered as an EDKII compatible protocol.

The underlying datum for the `Service` has a static lifetime, which
allows other protocols to consume it when producing their own services.
Due to this, all services are read-only; that is to say that only the
`&self` methods are available. If a service needs to be mutable, it
should manage it's own internal state via some sort of interior
mutability.

### Examples

``` rust
use uefi_sdk::component::params::Service;
use uefi_sdk::component::params::IntoService;
use uefi_sdk::component::Storage;
use uefi_sdk::error::Result;

trait MyService {
  fn do_something(&self) -> u32;
}

trait MyService2 {
  fn do_something_else(&self) -> u32;
}

#[derive(IntoService)]
#[service(dyn MyService)]
struct MyServiceImpl;

impl MyService for MyServiceImpl {
  fn do_something(&self) -> u32 {
    42
  }
 }

#[derive(IntoService)]
#[service(dyn MyService2)]
struct MyService2Impl {
  service: Service<dyn MyService>,
}

impl MyService2 for MyService2Impl {
  fn do_something_else(&self) -> u32 {
    self.service.do_something() + 1
  }
}

fn my_component(mut storage: Storage, service: Service<dyn MyService>) -> Result<()> {
    let s2 = MyService2Impl { service };
    storage.add_service(s2);
    Ok(())
}

#[derive(IntoService)]
#[service(dyn MyService, dyn MyService2)]
#[protocol = "cedc64fe-37f2-477d-9b01-7872dfe94ece"]
struct MyStruct;

```

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

100% test coverage

## Integration Instructions

N/A