rust-tool-cache: bugfix `rustup toolchain install` step (#282)
## Description

if rustup < 1.28.0, the `rustup toolchain install` command will fail,
but the cargo will automatically install the toolchain in the next step.
This is because rustup less than 1.28.0 automatically installs the
toolchain if it is not found whereas this functionality was removed in
1.28.0 in favor of being able to call rustup toolchain install without
the need to specify the toolchain name.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Via pipeline execution

## Integration Instructions

N/A