Add Driver Binding Abstraction (#33)
## Description

This PR introduces the driver binding abstraction to the UEFI SDK.

The driver binding abstraction consists of a `DriverBinding` trait used
to define the driver binding functionality of a driver using an
idiomatic rust approach and Types. Then a struct name
`UefiDriverBinding` is used to take ownership of any struct implementing
the `DriverBindign` to install and uninstall the driver binding by
managing the lifetime of the driver binding and the call to boot
services to install and uninstall the driver binding.


- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [X] Includes tests?
- [ ] Includes documentation?

## How Has This Been Tested?
- Added unit tests for the new driver_binding crate.
- Used this driver binding in UefiHidDxeV2 to verify the functionality.

## Integration Instructions

N/A