Ensure that page_pool doesn't grow unnecessarily (#285)
## Description

Observed that over time the page_pool continues to grow (i.e. not all
the pages allocated into the page_pool are always consumed). This patch
adds a check to ensure that the page_pool only increases by the amount
actually required, accounting for any unused pages that are already in
the pool.

It's anticipated this will be a short-term fix until the self-mapped
paging implementation is completed and tested.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

No longer observing FsbLock re-entrancy and winloader memory map
stability crashes with this modification.

## Integration Instructions

N/A