Add a function for creating a Result from efi::Status (#40)
## Description

Create a simple static function for converting a `efi::Status` to a
`error::Result`. This is useful when calling external C or Rust code
that returns a status as it's best to convert it to a Result for better
rust coding patterns. This is needed to reduce the use of efi::status in
other code locations that may depend on such external code.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

N/A