repository: Use local stacktrace when developing locally (#88)
## Description

With the previous setup, stacktrace was being pulled in via a registry
instead of using the local version. You would need to add a patch
statement to use the local version. This change updates the cargo.toml
so that the local version is used by default.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

N/A