Remove const-random crate (#302)
## Description

As specified in #296, the goal of removing `ahash` was actually to
remove the usage of the `const-random` (more specifically `tiny-keecak`)
crate. However, #301 immediately re-used the const-random crate to
generate the seed. This commit updates the seed to be the current OS
time.

To remind, this does not need to be cryptographically secure. This is
simply to enforce opacity regarding what the handle value will be so
that consumers cannot make assumptions based on the handle values.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Boot to UEFI Shell for Q35.

## Integration Instructions

N/A