spin_locked_gcd: Only Query First Page for Attributes
When deciding whether to map or remap a region, the GCD today
queries the entire region to decide which operation to do and if
remapping, which caching attributes to do. However, this is a
wasted operation, because if the region is unmapped, we attempt to
map and if this fails, it will bubble up a failure to us which is
the same as we would do when failing the query. If it is a remap
operation, the paging code does a query internally to ensure it
can remap and then does the remap if the range is consistent. In
the case of failure, remap will fail the same as if our original
query failed and if it succeeds, then the range is consistent
and we only need the cache attributes from the first entry in the
range to persist.

This patch updates to only query the first page and update the
logging in case of a paging error. This saves an unnecessary
page table walk.