spin_locked_gcd: Unmap Memory on remove_memory_space
When a memory space is removed from the GCD, we need to unmap
the pages in the page table, because DXE Core expects we will
not be using this any longer and we cannot rely on a free operation
to have done that, because other memory types such as MMIO and
reserved which are not allocated/freed can be removed here.

Platforms may also then try to add the memory space back later and
map it, which can fail if the region is already mapped.

This patch unmaps memory spaces when remove_memory_space is called
unconditionally, we expect it to fail if the memory region is
already mapped, but this is handled gracefully and execution
continues.