service_macro: Import storage struct from uefi_sdk::component (#55)
## Description

Fixes the compile error `cannot find ``Storage`` in this scope`. This
bug is because the macro is repsonsible for importing the necessary
modules / structs it uses, but failed to for the `Storage` struct,
relying on the user to import it themselves. This PR updates the code
that the macro generates to import the `Storage` struct from the
`uefi_sdk::component` module itself.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

All tests pass

## Integration Instructions

Users can stop importing the `Storage` struct themselves.