Add SMM Perf Records (#284)
## Description

Add smm performance record to the FPDT in an event running at ready to
boot.

Some abstraction for communication protocol has been added here. Will
get move eventually to a common place when we will be satisfied with
those.

- [x]  Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- Tested in QEMU
Add some random perf measurements in smm to be able to validate that we
can parse them in dxe core

- Tested on real hardware
Can boot to windows.

## Integration Instructions

N/A

# Note:
Some comments are in this pr: #271

---------

Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>