Add Code Organization doc (#334)
## Description

Closes #15 

The Rust DXE Core is a complex system with many constituent parts. This
document describes the organization of the overall codebase, including
the key dependencies that are shared between the Rust DXE Core and other
components. The goal is to provide a high-level overview of these
relationships.

This is meant to be a living document, and as the code base evolves,
this document should be updated to reflect the current state.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

## How This Was Tested

- `cargo make all`

## Integration Instructions

- N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>